Refactor: GameController와 Game 분리
